4 ran tests, no prediction, got red

+++ files/Hiker.kt
@@ -1,5 +1,5 @@
 package hiker
 
 fun minimum(arr:Array<Int>): Double{
-    return -2
+    return -2.toDouble()
 }
